Added new Titus Pod Identity Document
This is a new type of identity document for metatron
to use that will be "pod-aware".

This is a pretty big departure from what we used to do,
which was to provide a very titus'y document.

With this change, we just give away the whole pod and
expect the consumer to know how to understand it,
*and* to verify the pod "looks" correct based on what
it is expected.

I'm not 100% sure about the whole embedding of the pod
in terms of bytes there. The docs say we could either
just put the whole json in there or binary proto stuff,
but I do like the idea of not changing the api definitions
as the pod spec evolves. This puts a bit more burden on
the consumer, because they need to decode the pod themselves.