Added support for requesting platform sidecars on jobs
This is v0 of the platform sidecar stuff.

It includes the bare minimum that we will need (the name)
for including a platform sidecar.

I kinda hope this is all we will need! API stuff is a bit expensive,
and if we can get away with just saying "I want gandalf" then that would be
great.

"sidecar packs" are out of scope here of course, they come from a higher
level orchistrator. (newt may create your spin pipeline requesting
gandalf, metatron, shrimpi, etc for you)

This may feel a tad early, but because it takes so long for API requests
to get to the point where they are usable, I wanted to get a head start.