Added ContainerState to the TaskStatus message
As we move to a N-container world, we need some way to
report how *each* container is running.

This expands the Titus API to include the analogous API to the
k8s container status:
https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.19/#containerstatus-v1-core

We don't need everything here, I'm just getting started with
name and health status.

The *actually* k8s ContainerStatus via the pod object will still
be the transport for this data! It is just the control plane
needs a way to bubble it up, and eventually show it to things
like the Titus UI and Spinnaker.