Added support for specifying extraContainers in a job
We finally made it! We are ready to allow users to define
multiple containers in a Job.

In order to keep backwards compatibility, I left the 'main' container
message as-is. This simplifies a lot of stuff for us and our integrators.
The main container will still carry weight for things like the task's
resources, security stuff, attributes, and constraints.

At the same time, I want this feature as not-confusing as possible.
By calling it "extraContainers", it makes it more clear that one does
not need to duplicate the main container definition twice in the job
definition. These really are extra, and the main container stays.

This is not a complete representation of a container. It still needs room
for things like volumes, healthchecks, restart policies, etc. These
will be added in subsequent PRs.