RFC: Generate openapi (swagger) files from proto
It would be cool to bring back OpenAPI. I was wondering if there
was some programmatic way to go back and forth. Seems that there
is, kinda.

This uses this project:
https://grpc-ecosystem.github.io/grpc-gateway/

Which requires one to sprinkle in route annotations into the proto file.

Here is an example:
https://github.com/grpc-ecosystem/grpc-gateway/blob/master/examples/internal/proto/examplepb/echo_service.proto#L56

In this PR I only added in the Terminate Jobs route + DELETE, but of course
we can add more over time.