Turned MountPropagation into an enum instead of a string
In the k8s API, this is a string, but I thought we could make
it a little easier for our users and restrict what they
can choose.

Additionally, they won't have to look up the exact string they want,
because from the k8s docs it wasn't clear if you should put "None" or
"MountPropagationNone". No confusion with an enum though.