Removed swagger-v2 schemas
I would like to officially remove titus-v2 swagger support.
Surely nobody is using this, right?

We are pretty close to having auto-generated swagger for the v3 api
working again.

Once I get swagger on the v3 api working, I would like to add
a gradle task to scrape them and commit them in here too, to
make them easier to see, even if they are not the source of truth.