Added new HighScalability network mode
This change adds a new networking mode to the list of available modes
for our API.

The new mode is an *opinionated* way of configuring networking, which
*removes* options like subnets and security groups (maybe more in the
future).

One might think that this could be done via the spinnaker UI, but that
would make it harder for our direct-api users (non-spinnaker) to use
this. The theory here is that we (titus) want to be able to change the
semantics for what this mode implies for *all* workloads that request
it, not just ones launched via spinnaker.