Added Image data to ContainerState
Now that we have ContainerState working end-to-end, and actually visible
in the titus UI (and soon spinnaker UI), I would like to add Image data.

Just knowing the name of the container is a good start (it helps power
easy ssh links), but it would be also nice to know exactly what image
was used.

This this may be easy to see if you cross referenced it with a job, but
for platform-sidecars, the decision of exactly which version to use is
left to the runtime (webhook).

This PR allows us to get that data back out to the user in a structured
way.