Added platformSidecar to the container status message
When we get the container statuses, we can't be 100% sure about which
containers have been added by which platform sidecar. Sure we could
inspect the job and "subtract" the user-containers, but still we
wouldn't be sure which platform sidecar injected which.

This information *is* encoded in the lowest levels of the pod via
annotations, but I don't want to make spinnnaker and other tools have to
do the calculation.

This new change adds room to the proto for the titus API to simply just
say which platform sidecar added the container (because the API can see
the pod, and can do the logic).