RFC: Per-container capabilities
For a long time, we "hid" capabilities in container attributes.
I think we are ready to promote this feature as a first class citizen in
the API, and extend it to be per-container (not for a whole pod like
attributes are)

Also, I would like an ENUM so that we can package up related security
features (seccomp, apparmor, CAPs) into a high-level package, and not
expose individiual security settings to a user. (so you can't forget to
set the right fuse/seccomp/cap settings all correctly to enable fuse.